型を64bit符号付整数に変更

+++ src/consts_and_types.h
@@ -2,8 +2,7 @@
 
 #include <vector>
 
-using NumType = int;
+using NumType = long long;
 using SharesType = std::vector<NumType>;
 constexpr int PARTY_COUNT = 2;
 constexpr int MODULUS = 1 << 25;
-
